requirements.txt
# Este archivo solo contiene dependencias globales o de desarrollo para el proyecto Biblioperson.
# Para dependencias específicas, usa:
#   - backend/requirements.txt para el backend Flask
#   - dataset/requirements.txt para procesamiento de datos y dataset

# Ejemplo de dependencias globales (puedes agregar aquí herramientas de linting, testing, etc.)

# pip install -r backend/requirements.txt
# pip install -r dataset/requirements.txt

# ============================================
# DEPENDENCIAS CORE DE BIBLIOPERSON
# ============================================

# Dependencias para Biblioperson - Data Processing
ijson==3.2.1
python-dateutil==2.8.2
python-docx==1.0.1
pandas==2.1.0
openpyxl==3.1.2
PyMuPDF==1.23.3  # Reemplaza PyPDF2 para mejor rendimiento
textract==1.6.5 

# Dependencias para procesamiento semántico
sentence-transformers==2.2.2
numpy==1.24.3
spacy==3.5.3
tqdm==4.66.1 

# Dependencias para APIs de LLM
flask
flask-cors
python-dotenv
scikit-learn
google-generativeai
openai
anthropic
requests 

# Dependencias para GUI
PySide6==6.6.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
requests>=2.28.0
PyYAML>=6.0
aiohttp==3.12.0 

# ============================================
# SISTEMA OCR INTELIGENTE
# ============================================

# OCR Core - Tesseract y procesamiento de imágenes
pytesseract==0.3.10
Pillow==10.0.0

# PDF → Markdown Estructurado (pymupdf4llm)
pymupdf4llm==0.0.24

# OCR Cloud Providers (opcional - se instalan bajo demanda)
# Para Google Vision API
google-cloud-vision==3.4.0

# Para AWS Textract
boto3==1.26.0

# Para Azure Cognitive Services
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1

# Dependencias adicionales para procesamiento de imágenes
opencv-python-headless==4.8.0.76

# ============================================
# UTILIDADES DE DESARROLLO
# ============================================

# Detección de idiomas
langdetect==1.0.9

# Logging y monitoreo
coloredlogs==15.0.1

# Variables de entorno
python-decouple==3.8 